# [Testers Needed][Optimization] SPU/PPU: Qualify PUTLLC updates with no data changes