Please test if there are regressions albeit there should be none. This affects TSX and non-TSX as one.