@elad335 2022-05-18
Added another optimization mainly affecting TSX.