@Xcedf 2022-05-18
Perhaps MGO has smth speciffic, but have no fps difference in MGS4