@cipherxof 2022-05-18
> Perhaps MGO has smth speciffic, but have no fps difference in MGS4

That's definitely the case, but I'm not sure what. It's quire frustrating because we would like to move on from 0.0.8 to a more recent build. I know both engines are **very** similar. MGO is based on the MGS4 engine and they share much of the same code. I tried disabling all HLE network related functions but it had no effect on performance.