@cipherxof 2022-05-18
> Retest.

Performance is worse after [dfc9815](https://github.com/RPCS3/rpcs3/pull/12030/commits/dfc9815a3dde54d8da1657eec547c654e305c029)

<details>

  <summary>Latest Commit From this PR:</summary>

![Untitled](https://user-images.githubusercontent.com/5994581/169141841-6085a6fc-28db-43ed-beee-5ab670955d5c.png)

</details>

Once again, performance is significantly better / back to normal with Accurate RSX/PPU Reservations enabled.

I did notice the FPS in the game menus seems to be faster than before with this PR. I am reaching ~430fps in the menus whereas before it would always stay below 400.