@cipherxof 2022-05-18
Upon re-testing performance seems a lot closer. Maybe some variance due to exact camera location and the training dummies moving in the background.

<details>
  <summary>Re-test (with log)</summary>
  
[RPCS3.log.gz](https://github.com/RPCS3/rpcs3/files/8720429/RPCS3.log.gz)

![Untitled](https://user-images.githubusercontent.com/5994581/169144436-7a67ff27-3e74-46b5-9711-7cd51ed9ebce.png)

</details>