@AniLeo 2022-05-19
Steam Deck - PUTLLC

Before: 2408 ms
After: 545 ms