@elad335 2022-05-19
Added an ultra hack for non TSX, please test if improves or regresses performance in games.