@Xcedf 2022-05-19
Last commit worse in the test
Before:
SPU MFC 64 Bits War completed in 4535 ms (PS3: 3370 ms)
PPU/SPU MFC 64 Bits War completed in 7171 ms (PS3: 4443 ms)
After:
SPU MFC 64 Bits War completed in 5623 ms (PS3: 3370 ms)
PPU/SPU MFC 64 Bits War completed in 8081 ms (PS3: 4443 ms)