@cipherxof 2022-05-19
> Added an ultra hack for non TSX, please test if improves or regresses performance in games.

Performance is **significantly** better now in MGO. With Accurate PPU Reservations disabled, performance actually exceeds what it did before.

![image](https://user-images.githubusercontent.com/5994581/169248222-1583897b-87a9-4081-bf28-3c70f3d07d2b.png)