@elad335 2022-05-19
Updated the hack to be near purely performance advantageous.