@cipherxof 2022-05-19
> Updated the hack to be near purely performance advantageous.

I'm seeing a few percent increase in fps in some areas. In other areas it's about the same. In any case, this hack is a massive improvement for MGO. I also tested MGS4 briefly, and I can't say I noticed any difference on my machine. I suspect it may improve performance on machines that are SPU bottlenecked though.