@raveskirza 2022-05-19
In "Beyond: Two souls" the performance in the Intro screen has reduced

Master : 8-10 FPS 
[RPCS3_Master.log.gz](https://github.com/RPCS3/rpcs3/files/8728434/RPCS3_Master.log.gz)

PR: 5-9 FPS 
[RPCS3_PR.log.gz](https://github.com/RPCS3/rpcs3/files/8728436/RPCS3_PR.log.gz)

Note: This is on a quad Core CPU, and the game uses SPU heavily & gets bottlenecked due to the low CPU core count.