@cipherxof 2022-05-19
I'm reporting back after some more testing with Metal Gear Online. I was curious how it would run for lower end hardware, and for people that would struggle to maintain 30fps previously. I also tested on my machine under Linux, and performance even better there.

<details>
  <summary>Linux 12700k (+20fps)</summary>
  
On Windows this same spot is 118fps. Without this PR (on Windows) it's 108fps.

![unknown](https://user-images.githubusercontent.com/5994581/169384905-3359abde-e2cf-497e-a982-7007d2670659.png)
  
</details>

The following were taken on a i7 10700, but with a severely gimped off-brand motherboard and power supply. This user in particular struggled to get decent fps. Even my old r5 3600 was performing much better his CPU. It's hard to capture direct comparisons in a multiplayer match, so I also included the gameplay streams.

<details>
  <summary>Before this PR (10700):</summary>

Full video: https://youtu.be/GcKcCkcuvhg?t=7491
![Screenshot from 2022-05-19 12-17-06](https://user-images.githubusercontent.com/5994581/169385501-6706a9c0-118b-4ff9-b88c-1c351c555e43.png)

![20220517_163954](https://user-images.githubusercontent.com/5994581/169386936-f57423dc-fc35-4c13-8b17-7743ac659fa9.jpg)
</details>

<details>
  <summary>PR (10700):</summary>

Full video: https://youtu.be/r31w0TJzqNg

![Screenshot from 2022-05-19 12-24-38](https://user-images.githubusercontent.com/5994581/169386551-08d56ed1-c2b9-4b03-9d63-1bc531b9e932.png)

![20220519_153106](https://user-images.githubusercontent.com/5994581/169386765-5a12a3ce-dc08-4168-a91c-5c880a5dc7ba.jpg)


</details>

The biggest difference is in the lows. I'm guessing this is mainly due to not need Accurate PPU Reservations anymore. We essentially saw improvements across the board. We tested with a 3600, 3700x, 5950x, 9700k, 10700, and a 12700k.