@cipherxof 2022-05-20
> What do you mean not needing Accurate PPU Reservations? this pr is not meant to fix that.

I mean that's probably the main reasons FPS is so much better in MGO, whereas other titles that didn't require the setting are only seeing marginal improvements.