@Nerboruto 2022-08-21
is it possible to retest this with the latest optimizations introduced?
or is it obsolete?